More cleanup and some helpful comments (hopefully)
