+++ src/org/rascalmpl/library/util/ParseErrorRecovery.rsc
@@ -76,27 +76,4 @@ Tree(Tree) createParseErrorFilter(bool allowAmbiguity) =
 
 
 @javaClass{org.rascalmpl.library.util.ParseErrorRecovery}
-java &T<:Tree disambParseErrorsNoAmbiguity(&T<:Tree t);
-
-@javaClass{org.rascalmpl.library.util.ParseErrorRecovery}
 java &T<:Tree pruneAmbiguities(&T<:Tree t, int maxDepth=3);
-
-@javaClass{org.rascalmpl.library.util.ParseErrorRecovery}
-@synopsis{"curry" just the keyword parameters in a function with one argument.}
-@reflect{Needs to create a function}
-java &T () curry(&T () func, map[str, value] keywordParams=());
-
-@javaClass{org.rascalmpl.library.util.ParseErrorRecovery}
-@synopsis{"curry" just the keyword parameters in a function with one argument.}
-@reflect{Needs to create a function}
-java &T (&A param) curry(&T (&A param) func, map[str, value] keywordParams=());
-
-@javaClass{org.rascalmpl.library.util.ParseErrorRecovery}
-@synopsis{Curry a single argument function.}
-@reflect{Needs to create a function}
-java &T () curry(&T (&A param) func, &A arg, map[str, value] keywordParams=());
-
-@javaClass{org.rascalmpl.library.util.ParseErrorRecovery}
-@synopsis{Curry a two argument function.}
-@reflect{Needs to create a function}
-java &T (&A param) curry(&T (&A param1, &B param2) func, &B arg2, map[str, value] keywordParams=());
